Fix script server
Our new inventory isn't going to define ansible_host for everything (and
instead rely on dns/the hosts file), so we need our script server config
to be fixed up properly.

This change makes it so we we lookup the hostname to find the ip, and if
that fails, we fall back to an ansible_host entry.